AI_Football_Scout_Predictor/requirements.txt
requests>=2.31.0
pandas>=2.1.0
playwright>=1.40.0
playwright-stealth>=1.0.6
beautifulsoup4>=4.12.2
ollama>=0.2.0
python-telegram-bot>=21.0
